Refactor builder pattern
