+++ README.md
@@ -30,15 +30,14 @@ A basic analysis can be executed with the following example:
 ```java
 public class Main {
   public static void main(String[] args) {
-      DataFlowConfidentialityAnalysis analysis = new DataFlowAnalysisBuilder()
-          .standalone()
-          .modelProjectName("<PROJECT_NAME>")
-          .useBuilder(new PCMDataFlowConfidentialityAnalysisBuilder())
-          .usePluginActivator(Activator.class)
-          .useUsageModel("<USAGE_MODEL_PATH>")
-          .useAllocationModel("<ALLOCATION_MODEL_PATH>")
-          .useNodeCharacteristicsModel("<NODE_MODEL_PATH>")
-          .build();
+      DataFlowConfidentialityAnalysis analysis = new PCMDataFlowConfidentialityAnalysisBuilder()
+	        .standalone()
+	        .modelProjectName("<PROJECT_NAME>")
+	        .usePluginActivator(Activator.class)
+	        .useUsageModel("<USAGE_MODEL_PATH>")
+	        .useAllocationModel("<ALLOCATION_MODEL_PATH>")
+	        .useNodeCharacteristicsModel("<NODE_MODEL_PATH>")
+	        .build(); 
 
       analysis.setLoggerLevel(Level.TRACE); // Set desired logger level. Level.TRACE provides additional propagation Information
       analysis.initializeAnalysis();
