+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/StandalonePCMDataFlowConfidentialityAnalysis.java
@@ -7,7 +7,6 @@ import java.util.stream.Collectors;
 
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
-import org.dataflowanalysis.analysis.builder.AnalysisData;
 import org.dataflowanalysis.analysis.core.AbstractActionSequenceElement;
 import org.dataflowanalysis.analysis.core.ActionSequence;
 import org.dataflowanalysis.analysis.core.ActionSequenceFinder;
