+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/core/dfd/DFDActionSequenceElement.java
@@ -6,7 +6,7 @@ import org.eclipse.emf.ecore.EObject;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.dataflowanalysis.analysis.builder.AnalysisData;
+import org.dataflowanalysis.analysis.AnalysisData;
 import org.dataflowanalysis.analysis.core.AbstractActionSequenceElement;
 import org.dataflowanalysis.analysis.core.CharacteristicValue;
 import org.dataflowanalysis.analysis.core.DataFlowVariable;
