+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/core/pcm/user/CallingUserActionSequenceElement.java
@@ -3,7 +3,7 @@ package org.dataflowanalysis.analysis.core.pcm.user;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.dataflowanalysis.analysis.builder.AnalysisData;
+import org.dataflowanalysis.analysis.AnalysisData;
 import org.dataflowanalysis.analysis.core.AbstractActionSequenceElement;
 import org.dataflowanalysis.analysis.core.CharacteristicValue;
 import org.dataflowanalysis.analysis.core.DataFlowVariable;
